requirements.txt
asgiref==3.8.1
Django==5.2
django-allauth==65.7.0
django-cotton==2.0.3
django-extensions==3.2.3
django-filter==25.1
django-htmx==1.23.0
django-vite==3.1.0
factory_boy==3.3.3
Faker==37.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
